+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+* Correctly handle password retries when entering an invalid password on the
+  official Bitwarden server.
+
 ## [0.3.5] - 2020-05-25
 
 ### Fixed